deps/v8/test/webkit/fast/js/Promise-onFulfilled-deep-expected.txt

Test whether deeply chained then-s work.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is undefined
PASS result is 5042
PASS successfullyParsed is true

TEST COMPLETE
